Add an option to keep dw1000 off while processing incoming packet
This reduces the performance as the radio is kept off for longer time
but enables the application read all the registers (RX diagnostics, CIR, etc)
for the incoming packet right in the rx callback of Contiki stack.

I also removed the automatic reading of RX diag from dw1000.c, as this
can be done in a more flexible way by the application.